4633: process nits steven
